# Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoPulse API",
    "version": "1.0.0",
    "description": "API for managing driving license uploads and operations"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/fuel": {
      "post": {
        "summary": "Create a new fuel log",
        "tags": [
          "Fuel"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bike",
                  "date",
                  "amount",
                  "unitCost",
                  "odometer"
                ],
                "properties": {
                  "bike": {
                    "type": "string",
                    "description": "Bike ID",
                    "example": "685d4f07cc39dcbd6a0171a0"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-06-27"
                  },
                  "amount": {
                    "type": "number",
                    "example": 10
                  },
                  "unitCost": {
                    "type": "number",
                    "example": 120
                  },
                  "odometer": {
                    "type": "number",
                    "example": 9600
                  },
                  "note": {
                    "type": "string",
                    "example": "Full tank"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fuel log saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Fuel log saved successfully"
                    },
                    "fuelLog": {
                      "$ref": "#/components/schemas/FuelLog"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Bike not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/fuel/{bikeId}": {
      "get": {
        "summary": "Get all fuel logs for a specific bike",
        "tags": [
          "Fuel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bikeId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bike ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of fuel logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuelLog"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/fuel/{fuelLogId}": {
      "delete": {
        "summary": "Delete a fuel log by ID",
        "tags": [
          "Fuel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fuelLogId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Fuel log ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Fuel log deleted"
          },
          "404": {
            "description": "Fuel log not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a fuel log by ID",
        "tags": [
          "Fuel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fuelLogId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Fuel log ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "example": 12
                  },
                  "unitCost": {
                    "type": "number",
                    "example": 125
                  },
                  "odometer": {
                    "type": "number",
                    "example": 9700
                  },
                  "note": {
                    "type": "string",
                    "example": "Partial fill"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fuel log updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelLog"
                }
              }
            }
          },
          "404": {
            "description": "Fuel log not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}